blog_generator/requirements.txt
flask==3.0.0
google-generativeai==0.3.2
markdown==3.5.1
